pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "regulatory-gap-analyzer"
version = "0.1.0"
description = "A compliance research tool for identifying regulatory ambiguity and divergence"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "RegulatoryGapAnalyzer Team"}
]
keywords = [
    "regulatory",
    "compliance",
    "gap-analysis",
    "legal",
    "fintech",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Legal Industry",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "numpy>=1.21.0",
]

[project.optional-dependencies]
pdf = ["pypdf>=3.0.0"]
html = ["beautifulsoup4>=4.11.0"]
docx = ["python-docx>=0.8.11"]
all = [
    "pypdf>=3.0.0",
    "beautifulsoup4>=4.11.0",
    "python-docx>=0.8.11",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
reg-gap = "reg_gap.cli:main"

[project.urls]
Homepage = "https://github.com/example/regulatory-gap-analyzer"
Documentation = "https://github.com/example/regulatory-gap-analyzer#readme"
Repository = "https://github.com/example/regulatory-gap-analyzer"

[tool.setuptools.packages.find]
where = ["."]
include = ["reg_gap*"]

[tool.pytest.ini_options]
testpaths = ["reg_gap/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
